Major cleans to being coding again
- Add requirements.txt
- Update documentation (including reformatting to MD)
- Make sure all tests run again
- Remove code in rsmtpd/__init__.py
