+++ tests/handlers/test_reject_all.py
@@ -1,13 +1,13 @@
 from logging import Logger
-from unittest import TestCase
-
-from rsmtpd import ConfigLoader, LoggerFactory
+from rsmtpd.core.config_loader import ConfigLoader
+from rsmtpd.core.logger_factory import LoggerFactory
 from rsmtpd.handlers.reject_all import RejectAll
 from rsmtpd.handlers.shared_state import SharedState
 from rsmtpd.response.action import CLOSE, OK, CONTINUE
 from rsmtpd.response.smtp_354 import SmtpResponse354
 from rsmtpd.response.smtp_521 import SmtpResponse521
 from tests.mocks import MockConfigLoader, StubLoggerFactory
+from unittest import TestCase
 
 
 class TestRejectAllHandler(TestCase):
